feat: base yml parse

+++ data-files/src/domain/transformation.rs
@@ -1,13 +1,13 @@
 use super::character::Character;
 use std::fmt::Debug;
 
-#[derive(Debug)]
+#[derive(Debug, PartialEq, Eq)]
 pub enum Operation {
     Add(u8),
-    CeilFrac(f32),
+    CeilFrac(u8),
 }
 
-#[derive(Debug)]
+#[derive(Debug, PartialEq, Eq)]
 pub enum Property {
     Constitution,
     Willpower,
@@ -24,7 +24,7 @@ impl Transformation {
     pub fn transform(&self, c: Character) -> Character {
         let closure: Box<dyn Fn(u8) -> u8> = match self.operation {
             Operation::CeilFrac(divider) => {
-                Box::new(move |nb: u8| f32::ceil(nb as f32 / divider) as u8)
+                Box::new(move |nb: u8| f32::ceil(nb as f32 / divider as f32) as u8)
             }
             Operation::Add(addition) => Box::new(move |nb: u8| nb + addition),
         };
